legacy/requirements.txt
selenium>=4.3.0
unidecode>=1.3.4
python-dotenv
google
html2text
scipy
tenacity
tiktoken
termcolor 
openai
requests
favicon
Pillow
psycopg2
lxml
bs4
tqdm
openpyxl 